+++ app-backend/src/__tests__/report-service/resolver.test.ts
@@ -10,7 +10,6 @@ jest.mock('../../report-service/models/Report');
 
 import {ReportAttributes, Report} from '../../report-service/models/Report';
 import {reportResolver, CreateOrUpdateAttributes} from '../../report-service/api/resolver';
-import {userDataMapResolver} from '../../user-data-map/api/resolver';
 import {genotypeResolver} from '../../genotype-service/api/resolver';
 
 const unroll = require('unroll');
@@ -29,36 +28,25 @@ describe('Report resolver tests.', () => {
   };
 
   const dummyRequestData: CreateOrUpdateAttributes = {...commonData, rawContent: 'demo-content'};
-  const dummyResponseData: ReportAttributes = {...commonData, raw_content: 'demo-content'};
+  const dummyResponseData: ReportAttributes = {...commonData, rawContent: 'demo-content'};
 
   Report.createAsync = jest.fn()
-      .mockImplementation((data: ReportAttributes): {attrs: ReportAttributes} => ({attrs: data}))
+      .mockImplementation((data: ReportAttributes) => ({
+        get: (): ReportAttributes => data
+      }))
       .mockImplementationOnce(() => {throw new Error('createAsync mock error'); });
 
-  const mockedExecAsync: jest.Mock<ExecSuccess> = jest.fn((): {execAsync: jest.Mock<ExecSuccess>} => {
-    return {
-      execAsync: jest.fn((): ExecSuccess => ({Items: [dummyResponseData]}))
-    };
-  });
-
-  const mockedLimit = jest.fn((limit: number) => ({
-    execAsync: jest.fn((): ExecSuccess => ({Items: [dummyResponseData]})),
-    startKey: mockedExecAsync,
-  }));
+  Report.getAsync = jest.fn()
+      .mockImplementation((report, slug) => (slug && {get: (): ReportAttributes => dummyResponseData}))
+      .mockImplementationOnce(() => {throw new Error('getAsync mock error'); });
 
   genotypeResolver.list = jest.fn();
-  
-  Report.query = jest.fn(() => {
-    return {
-      limit: mockedLimit,
-      usingIndex: jest.fn(() => ({limit: mockedLimit}))
-    };
-  })
-  .mockImplementationOnce(() => {throw new Error('query mock error'); });
 
-  userDataMapResolver.get = jest.fn()
-      .mockImplementation((data: ReportAttributes): {attrs: ReportAttributes} => ({attrs: data}))
-      .mockImplementationOnce(() => {throw new Error('userDataMapResolver mock error'); });
+  Report.query = jest.fn()
+    .mockImplementationOnce(() => {throw new Error('query mock error'); })
+    .mockImplementation(() => ({
+      execAsync: jest.fn(() => ({Items: [{get: () => ({...commonData, rawContent: 'demo-content'})}]}))
+    })
 
   describe('Create tests', () => {
     it('should throw an error when the record already exists.', async () => {
@@ -75,51 +63,45 @@ describe('Report resolver tests.', () => {
   describe('List test', () => {
 
     it('should fail if an error occurs', async () => {
-      let response = await reportResolver.list({}, authorizer);
+      let response = await reportResolver.list({});
       expect(response[`message`]).toEqual('query mock error');
     });
-  
+
     unroll('it should respond with the report data list when #params are present.', async (
         done: () => void,
         args: {params: {[key: string]: string | number}}
     ) => {
-      let response = await reportResolver.list(args.params, authorizer);
-      expect(response).toEqual({Items: [dummyResponseData]});
+      let response = await reportResolver.list(args.params);
+      expect(response).toEqual([dummyResponseData]);
       done();
     }, [ // tslint:disable-next-line
       ['params'],
-      [{slug: 'test', limit: 10}],
-      [{slug: 'QWERTY2', lastEvaluatedKeys: {slug: 'test', id: 'test'}}],
+      [{slug: 'test'}],
+      [{slug: 'QWERTY2'}],
     ]);
   });
 
   describe('Get tests', () => {
     it('should throw an error when the if user is not found.', async () => {
       let response = await reportResolver.get({}, authorizer);
-      expect(response[`message`]).toEqual('userDataMapResolver mock error');
+      expect(response[`message`]).toEqual('getAsync mock error');
     });
 
     it('should return an error message if required parameters are not present.', async () => {
       let response = await reportResolver.get({}, authorizer);
-      expect(response[`message`]).toEqual('Required parameters not present.');
+      expect(response[`message`]).toEqual('No such record found');
     });
 
-    unroll('it should respond with the report data list when #params are present.', async (
-        done: () => void,
-        args: {params: {[key: string]: string | number}}
-    ) => {
-      let response = await reportResolver.get(args.params, authorizer);
-      response[`userData`]({});
-      expect(genotypeResolver.list).toBeCalled();
-      expect(response[`Items`]).toEqual([dummyResponseData]);
-      expect(typeof response[`userData`]).toEqual('function');
-      done();
-    }, [ // tslint:disable-next-line
-      ['params'],
-      [{id: 'test'}],
-      [{slug: 'test'}],
-      [{slug: 'test', lastEvaluatedKeys: {slug: 'test', id: 'test'}}],
-    ]);
+    it('should return data successfully when no error occurs', async () => {
+      let response = await reportResolver.get({slug: 'test'}, {claims: {sub: 'demo-id'}});
+      let userData = response.userData();
+
+      expect(userData.genotypes).toBeInstanceOf(Promise);
+      expect(response[`userData`]).toBeInstanceOf(Function);
+      expect(response).toHaveProperty('title', 'demo-title');
+      expect(response).toHaveProperty('slug', 'demo-slug');
+      expect(response).toHaveProperty('genes', [ 'demo', 'genes' ]);
+    });
   });
 
 });